The zoom is now ok, the color scheme has some issues but it’s no really a big deal. The code will have to be cleaned!

+++ include/mandelbrot.h
@@ -4,13 +4,13 @@
 #include <stdbool.h>
 
 struct FR_Mandelbrot_Complex {
-  double r;
-  double i;
+  long double r;
+  long double i;
 };
 
 typedef struct FR_Mandelbrot_Complex FR_Mandelbrot_Complex;
 
-void FR_Mandelbrot_uv2mandelbrot(double *x, double *y); 
+void FR_Mandelbrot_uv2mandelbrot(long double *x, long double *y); 
 void FR_Mandelbrot_next(FR_Mandelbrot_Complex *zn, FR_Mandelbrot_Complex const *c);
 int FR_Mandelbrot_eval(FR_Mandelbrot_Complex const *xy, int maxi);
